+++ gui/extension/CHANGELOG.md
@@ -9,6 +9,7 @@
 ### Fixes
 
 - BUG#37697639 FK violation regarding derived role on DROP REST SERVICE
+- BUG#38461981 fixed point decimals losing precision in the python sdk
 -
 
 ## Changes in 1.19.17+9.4.0